Extend library to support PCF8575 as well.
The 75 device is exactly the same as the 74 device, except that the
former has 16 GPIO pins, and is read from and written to in 16 bit words
instead of 8 bit words.

Rename the whole thing to pcf857x and add two pcf8575_create() and
pcf8574_create() functions. The only difference is the setting of
dev->num_gpios.